+++ docs/js/main.js
@@ -62,6 +62,9 @@ $(function () {
       },
       'zoomout.cropper': function (e) {
         console.log(e.type);
+      },
+      'change.cropper': function (e) {
+        console.log(e.type);
       }
     }).cropper(options);
 
@@ -71,6 +74,10 @@ $(function () {
       var data = $(this).data(),
           $target,
           result;
+
+      if (!$image.data('cropper')) {
+        return;
+      }
 
       if (data.method) {
         data = $.extend({}, data); // Clone a new one
@@ -103,6 +110,10 @@ $(function () {
 
       }
     }).on('keydown', function (e) {
+
+      if (!$image.data('cropper')) {
+        return;
+      }
 
       if (this.scrollTop > 300) {
         return;
@@ -143,6 +154,10 @@ $(function () {
         var files = this.files,
             file;
 
+        if (!$image.data('cropper')) {
+          return;
+        }
+
         if (files && files.length) {
           file = files[0];
 
@@ -165,6 +180,10 @@ $(function () {
     // Options
     $('.docs-options :checkbox').on('change', function () {
       var $this = $(this);
+
+      if (!$image.data('cropper')) {
+        return;
+      }
 
       options[$this.val()] = $this.prop('checked');
       $image.cropper('destroy').cropper(options);
@@ -189,8 +208,8 @@ $(function () {
       guides: false,
       highlight: false,
       dragCrop: false,
-      movable: false,
-      resizable: false
+      cropBoxMovable: false,
+      cropBoxResizable: false
     });
   })();